+++ src/postgres/teams/invitation.rs
@@ -18,12 +18,12 @@ impl PostgresTeamInvitationRepository {
 
 #[derive(FromRow)]
 struct InvitationRecord {
-    id: i32,
-    team_id: i32,
+    id: i64,
+    team_id: i64,
     email: String,
     role: String,
     token_hash: String,
-    invited_by: i32,
+    invited_by: i64,
     expires_at: DateTime<Utc>,
     accepted_at: Option<DateTime<Utc>>,
     created_at: DateTime<Utc>,
@@ -73,7 +73,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_id(&self, id: i32) -> Result<Option<TeamInvitation>, AuthError> {
+    async fn find_by_id(&self, id: i64) -> Result<Option<TeamInvitation>, AuthError> {
         let row: Option<InvitationRecord> = sqlx::query_as(
             "SELECT id, team_id, email, role, token_hash, invited_by, expires_at, accepted_at, created_at FROM team_invitations WHERE id = $1",
         )
@@ -108,7 +108,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_pending_by_team(&self, team_id: i32) -> Result<Vec<TeamInvitation>, AuthError> {
+    async fn find_pending_by_team(&self, team_id: i64) -> Result<Vec<TeamInvitation>, AuthError> {
         let rows: Vec<InvitationRecord> = sqlx::query_as(
             r"
             SELECT id, team_id, email, role, token_hash, invited_by, expires_at, accepted_at, created_at
@@ -150,7 +150,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn mark_accepted(&self, id: i32) -> Result<TeamInvitation, AuthError> {
+    async fn mark_accepted(&self, id: i64) -> Result<TeamInvitation, AuthError> {
         let row: InvitationRecord = sqlx::query_as(
             r"
             UPDATE team_invitations SET accepted_at = NOW()
@@ -173,7 +173,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete(&self, id: i32) -> Result<(), AuthError> {
+    async fn delete(&self, id: i64) -> Result<(), AuthError> {
         sqlx::query("DELETE FROM team_invitations WHERE id = $1")
             .bind(id)
             .execute(&self.pool)
@@ -187,7 +187,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete_expired(&self) -> Result<u64, AuthError> {
+    async fn delete_expired(&self) -> Result<i64, AuthError> {
         let result = sqlx::query("DELETE FROM team_invitations WHERE expires_at < NOW()")
             .execute(&self.pool)
             .await
@@ -196,6 +196,6 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
                 AuthError::DatabaseError(e.to_string())
             })?;
 
-        Ok(result.rows_affected())
+        Ok(result.rows_affected() as i64)
     }
 }
